revised README and fix Windows Int32 error

+++ src/AISIS10016.jl
@@ -125,7 +125,7 @@ function h121(Pbar, Mxbar, Mybar, Pa, Max, May)
 
 end
 
-function table23131(CorZ::Int64,t,b,d,θ)
+function table23131(CorZ::Int32,t,b,d,θ)
 
     Af=(b+d)*t
     Jf=1/3*b*t^3+1/3*d*t^3
